Least Squares Approximator

+++ leastsquares/matrixlib.h
@@ -15,11 +15,13 @@
  */
 
 #pragma once
+#include <sys/types.h>
 
-int solve_system(double *matrix, double *x, double *y, size_t *map, size_t n,
-                 size_t m);
+ssize_t solve_system(double *matrix, double *x, double *y, size_t *map,
+                     size_t n, size_t m);
 
-int qr_decompose(double *matrix, double *y, size_t *map, size_t n, size_t m);
+ssize_t qr_decompose(double *matrix, double *y, size_t *map, size_t n,
+                     size_t m);
 
 void gauss_back_substitute(double *matrix, double *x, double *y, size_t n,
 						   size_t m);
